Make SearchToolbar into a module
